usuarios/requirements.txt
fastapi==0.115.3
uvicorn==0.32.0

# Base de datos
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9

# Seguridad (JWT + hashing)
bcrypt==4.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4


# Validación y utilidades
pydantic==2.9.2

